caserver/api/requirements.txt
flask
pyjwt
uwsgi
mysql-connector-python
cryptography
